requirements.txt
streamlit==1.42.0
streamlit-reveal-slides==0.2.0
pandas==2.2.3
numpy==2.2.2
matplotlib==3.10.0
seaborn==0.13.2
scikit-learn==1.6.1